[android] move place page sections in own package
