+++ android/src/app/organicmaps/widget/placepage/sections/PlaceOpeningHoursAdapter.java
@@ -1,4 +1,4 @@
-package app.organicmaps.widget.placepage;
+package app.organicmaps.widget.placepage.sections;
 
 import android.view.LayoutInflater;
 import android.view.View;
