+++ android/src/app/organicmaps/widget/placepage/sections/PlacePagePhoneFragment.java
@@ -1,4 +1,4 @@
-package app.organicmaps.widget.placepage;
+package app.organicmaps.widget.placepage.sections;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,6 +14,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import app.organicmaps.R;
 import app.organicmaps.bookmarks.data.MapObject;
 import app.organicmaps.bookmarks.data.Metadata;
+import app.organicmaps.widget.placepage.PlacePageViewModel;
 
 public class PlacePagePhoneFragment extends Fragment implements Observer<MapObject>
 {
