+++ android/src/app/organicmaps/widget/placepage/sections/PlacePhoneAdapter.java
@@ -1,4 +1,4 @@
-package app.organicmaps.widget.placepage;
+package app.organicmaps.widget.placepage.sections;
 
 import android.content.Context;
 import android.text.TextUtils;
