+++ android/tests/java/app/organicmaps/widget/placepage/sections/PlaceOpeningHoursAdapterTest.java
@@ -1,9 +1,10 @@
-package app.organicmaps.widget.placepage;
+package app.organicmaps.widget.placepage.sections;
 
 import app.organicmaps.editor.data.HoursMinutes;
 import app.organicmaps.editor.data.Timespan;
 import app.organicmaps.editor.data.Timetable;
-import app.organicmaps.widget.placepage.PlaceOpeningHoursAdapter.WeekScheduleData;
+import app.organicmaps.widget.placepage.sections.PlaceOpeningHoursAdapter;
+import app.organicmaps.widget.placepage.sections.PlaceOpeningHoursAdapter.WeekScheduleData;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
